Lots of graphics work, as a result of running valgrind and finding that my shader code was ignoring some OpenGL errors (now it ignores fewer... ...). Switched from GLEE to GLEW for the extensions support past OpenGL 3.0, but my configuration doesn't currently support geometry shaders, so instead of figuring that out I just disabled them for now (they weren't going to be efficient in any case I think).
The switch from GLEE to GLEW made me overhaul my OpenGL checking file,
so it's now a real .c file instead of just a .h file to be compiled with
-E.

Also some progress on worldgen; slowly fixing bugs and speeding things
up there. Added a layer of B_BOUNDARY to the bottom of the world.

Also improved texture loading: if a texture fails to load, it doesn't
keep trying every frame.

-Peter Mawhorter